// src/types/api/PaginatedResponse.ts
export type PaginatedResponse<T> = {
    data: T[];
    links: {
        first: string | null;
        last: string | null;
        prev: string | null;
        next: string | null;
    }[];

    meta: {
        current_page: number;
        from: number | null;
        last_page: number;
        path: string;
        per_page: number;
        to: number | null;
        total: number;
        links: {
            url: string | null;
            label: string | null;
            active: boolean;
        }[];
    }
}